move cocuntry endpoints to new geostore service
